move type file
